split the commands out into separate files
Organize related groups together and top-level commands on their own.
