+++ src/fromager/commands/__init__.py
@@ -1,0 +1,8 @@
+from . import bootstrap, build_order, canonicalize, step
+
+commands = [
+    bootstrap.bootstrap,
+    build_order.build_order,
+    step.step,
+    canonicalize.canonicalize,
+]
