+++ tests/test_bootstrap.py
@@ -1,22 +1,22 @@
 import textwrap
 
-from fromager import __main__ as main
+from fromager.commands import bootstrap
 
 
 def test_get_requirements_single_arg():
-    requirements = main._get_requirements_from_args(['a'], [])
+    requirements = bootstrap._get_requirements_from_args(['a'], [])
     assert ['a'] == requirements
 
 
 def test_get_requirements_multiple_args():
-    requirements = main._get_requirements_from_args(['a', 'b'], [])
+    requirements = bootstrap._get_requirements_from_args(['a', 'b'], [])
     assert ['a', 'b'] == requirements
 
 
 def test_get_requirements_requirements_file(tmp_path):
     requirements_file = tmp_path / 'requirements.txt'
     requirements_file.write_text('c\n')
-    requirements = main._get_requirements_from_args([], [requirements_file])
+    requirements = bootstrap._get_requirements_from_args([], [requirements_file])
     assert ['c'] == requirements
 
 
@@ -30,7 +30,7 @@ def test_get_requirements_requirements_file_comments(tmp_path):
 
         '''),
     )
-    requirements = main._get_requirements_from_args([], [requirements_file])
+    requirements = bootstrap._get_requirements_from_args([], [requirements_file])
     assert ['c', 'd'] == requirements
 
 
@@ -39,19 +39,19 @@ def test_get_requirements_requirements_file_multiple(tmp_path):
     requirements1_file.write_text('a\n')
     requirements2_file = tmp_path / 'requirements2.txt'
     requirements2_file.write_text('b\n')
-    requirements = main._get_requirements_from_args([], [requirements1_file, requirements2_file])
+    requirements = bootstrap._get_requirements_from_args([], [requirements1_file, requirements2_file])
     assert ['a', 'b'] == requirements
 
 
 def test_get_requirements_file_with_comments_and_blanks(tmp_path):
     requirements_file = tmp_path / 'requirements.txt'
     requirements_file.write_text('a\n\n# ignore\nb\nc\n')
-    requirements = main._get_requirements_from_args([], [requirements_file])
+    requirements = bootstrap._get_requirements_from_args([], [requirements_file])
     assert ['a', 'b', 'c'] == requirements
 
 
 def test_get_requirements_args_and_file(tmp_path):
     requirements_file = tmp_path / 'requirements.txt'
     requirements_file.write_text('c\n')
-    requirements = main._get_requirements_from_args(['a', 'b'], [requirements_file])
+    requirements = bootstrap._get_requirements_from_args(['a', 'b'], [requirements_file])
     assert ['a', 'b', 'c'] == requirements
